+++ src/test/java/org/tikv/raw/RawKVClientTest.java
@@ -3,13 +3,11 @@ package org.tikv.raw;
 import static org.junit.Assert.*;
 
 import com.google.protobuf.ByteString;
-import java.io.IOException;
 import java.util.*;
 import java.util.concurrent.*;
 import java.util.stream.Collectors;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.After;
-import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.slf4j.Logger;
@@ -17,7 +15,6 @@ import org.slf4j.LoggerFactory;
 import org.tikv.common.TiConfiguration;
 import org.tikv.common.TiSession;
 import org.tikv.common.codec.KeyUtils;
-import org.tikv.common.exception.RawCASConflictException;
 import org.tikv.common.exception.TiKVException;
 import org.tikv.common.key.Key;
 import org.tikv.common.util.FastByteComparisons;
@@ -73,7 +70,7 @@ public class RawKVClientTest {
   }
 
   @Before
-  public void setup() throws IOException {
+  public void setup() {
     try {
       TiConfiguration conf = TiConfiguration.createRawDefault();
       session = TiSession.create(conf);
@@ -94,44 +91,6 @@ public class RawKVClientTest {
     if (session != null) {
       session.close();
     }
-  }
-
-  @Test
-  public void rawCASTest() {
-    if (!initialized) return;
-    ByteString key = ByteString.copyFromUtf8("key_atomic");
-    ByteString value = ByteString.copyFromUtf8("value");
-    ByteString value2 = ByteString.copyFromUtf8("value2");
-    client.delete(key);
-    client.compareAndSet(key, Optional.empty(), value);
-    Assert.assertEquals(value, client.get(key).get());
-    try {
-      client.compareAndSet(key, Optional.empty(), value2);
-      Assert.fail("compareAndSet should fail.");
-    } catch (RawCASConflictException err) {
-      Assert.assertEquals(value, err.getPrevValue().get());
-    }
-  }
-
-  @Test
-  public void rawPutIfAbsentTest() {
-    if (!initialized) return;
-    long ttl = 10;
-    ByteString key = ByteString.copyFromUtf8("key_atomic");
-    ByteString value = ByteString.copyFromUtf8("value");
-    ByteString value2 = ByteString.copyFromUtf8("value2");
-    client.delete(key);
-    Optional<ByteString> res1 = client.putIfAbsent(key, value, ttl);
-    assertFalse(res1.isPresent());
-    Optional<ByteString> res2 = client.putIfAbsent(key, value2, ttl);
-    assertEquals(res2.get(), value);
-    try {
-      Thread.sleep(ttl * 1000);
-    } catch (InterruptedException e) {
-      Thread.currentThread().interrupt();
-    }
-    Optional<ByteString> res3 = client.putIfAbsent(key, value, ttl);
-    assertFalse(res3.isPresent());
   }
 
   @Test
